bot/requirements.txt
aiogram>=3.0.0
redis>=4.5.0
httpx>=0.24.0
python-dotenv>=1.0.0
jinja2>=3.1.2
psycopg2-binary>=2.9.9
clickhouse-connect>=0.13.0
